requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.6
django-annoying==0.10.6
django-betterforms==1.2
django-betterforms3==2
django-debug-toolbar==3.5.0
django-probes==1.6.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==8.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pypyodbc==1.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
python-decouple
